VIDEO-521 | Refactor src/jwplayer/ - Styled Components to SCSS Modules
