+++ src/jwplayer/players/CanonicalVideoPlayer/CanonicalVideoPlayer.tsx
@@ -2,40 +2,25 @@ import React, { useEffect, useRef } from 'react';
 import { CanonicalVideoPlayerProps } from 'jwplayer/types';
 import PlayerWrapper from 'jwplayer/players/shared/PlayerWrapper';
 import LoadableVideoPlayerWrapper from 'jwplayer/players/shared/LoadableVideoPlayerWrapper';
-import styled, { css } from 'styled-components';
 import useAdComplete from 'jwplayer/utils/useAdComplete';
 import { communicationService } from 'jwplayer/utils/communication';
 import { isLocalDevelopment, isOnBrowser } from 'jwplayer/utils/envs';
 
-const CanonicalVideoTopPlaceholder = styled.div`
-	width: 100%;
-	position: relative;
-`;
+import clsx from 'clsx';
+
+import styles from './canonicalVideoPlayer.module.scss';
 
 interface CanonicalVideoWrapperProps {
 	isScrollPlayer: boolean;
 }
 
-const CanonicalVideoWrapper = styled.div<CanonicalVideoWrapperProps>`
-	@media only screen and (max-width: 767px) {
-		${(props) =>
-			props.isScrollPlayer
-				? css`
-						box-shadow: 0 2px 4px 0 rgb(0 0 0 / 20%);
-						position: fixed;
-						top: 55px;
-						width: 100%;
-						z-index: 10;
-				  `
-				: css`
-						transform: translateZ(0);
-						-webkit-transform: translateZ(0);
-						-webkit-transition: padding 0.3s;
-						transition: padding 0.3s;
-						padding: 0;
-				  `}
-	}
-`;
+const CanonicalVideoWrapper: React.FC<CanonicalVideoWrapperProps> = ({ isScrollPlayer }) => (
+	<div
+		className={clsx(
+			isScrollPlayer ? styles.canonicalVideoWrapperIsScrollPlayer : styles.canonicalVideoWrapperIsNotScrollPlayer,
+		)}
+	/>
+);
 
 const CanonicalVideoPlayer: React.FC<CanonicalVideoPlayerProps> = ({ currentVideo, videoDetails, onComplete }) => {
 	const ref = useRef<HTMLDivElement>(null);
@@ -53,7 +38,7 @@ const CanonicalVideoPlayer: React.FC<CanonicalVideoPlayerProps> = ({ currentVide
 
 	return (
 		<PlayerWrapper playerName="canonical-video-player">
-			<CanonicalVideoTopPlaceholder ref={ref}>
+			<div className={styles.canonicalVideoTopPlaceholder} ref={ref}>
 				{adComplete && (
 					<CanonicalVideoWrapper isScrollPlayer={false}>
 						<LoadableVideoPlayerWrapper
@@ -63,7 +48,7 @@ const CanonicalVideoPlayer: React.FC<CanonicalVideoPlayerProps> = ({ currentVide
 						/>
 					</CanonicalVideoWrapper>
 				)}
-			</CanonicalVideoTopPlaceholder>
+			</div>
 		</PlayerWrapper>
 	);
 };
